cloud_function_from_bq_to_cs.py/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery
os